refactor(analysis): extract function detail helpers
